+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## New
 - Added Hold mode in duty cycle since people couldn't figure out that 100% is the same as Holding down the mouse continuously.
 - Added slider for window opacity in the appearance settings.
+- Added Keybind section in settings.
+- Added Keybinds for switching between all pages.(1 to 5 for simple, advanced, zones, click points, settings)
 ## Changed
 - Overhauled Advanced panel.
 - Settings panel now has a side menu for easier navigation and organization of settings.